Separated out plotting functions into three separate scripts - storage_plots, file_plots and running_total_plots
