+++ Ploutos/scripts/storage_plots.py
@@ -4,335 +4,14 @@ import calendar
 import json
 import pandas as pd
 import plotly.express as px
-import plotly.graph_objects as go
 
-from collections import defaultdict
-from datetime import date, datetime
+from datetime import date
 from dateutil.relativedelta import relativedelta
-from more_itertools import unique_everseen
 
-from dashboard.models import StorageCosts, DailyOrgRunningTotal, FileTypeDate
+from dashboard.models import StorageCosts
 from django.conf import settings
 from django.db.models import Sum
 from scripts import DNAnexus_queries as dx_queries
-
-
-class RunningTotPlotFunctions():
-    """Class for plotting functions for the running total graph"""
-
-    def __init__(self) -> None:
-        self.today_date = dx_queries.no_of_days_in_month()[0]
-        self.four_months_ago = date.today() + relativedelta(months=-4)
-        self.start_of_four_months_ago = self.four_months_ago.replace(day=1)
-        self.start_of_next_month = (
-            date.today() + relativedelta(months=+1)
-        ).replace(day=1)
-
-        # Get all DailyOrgRunningTotal objects as a queryset
-        # So multiple db queries not needed
-        self.totals = DailyOrgRunningTotal.objects.all()
-
-    def calculate_diffs(self, tuple_of_charges):
-        """
-        Creates list of charge differences
-        Where one charge is subtracted from the previous date
-        Parameters
-        ----------
-        tuple_of_charges :  tuple
-            date-sorted tuple of charges for a specific charge type
-
-        Returns
-        -------
-        charge_diff : list
-            list with charge differences from the previous day
-        """
-        charge_diff = [
-            y - x for x, y in zip(tuple_of_charges, tuple_of_charges[1:])
-        ]
-
-        return charge_diff
-
-    def daily_plot(self, totals):
-        """
-        Set bar chart context for daily running charges plot
-        Parameters
-        ----------
-        totals :  queryset
-            queryset already filtered by default or specified daterange
-
-        Returns
-        -------
-        fig : Plotly figure object
-        """
-
-        # Order filtered DailyOrgRunningTotal objects by date and get charges
-        charges = totals.order_by(
-            'date__date'
-        ).values_list(
-            'storage_charges', 'compute_charges', 'egress_charges'
-        )
-
-        # Get dates in order from db
-        dates = totals.order_by('date__date').values_list(
-            'date__date', flat=True
-        )
-
-        # Turn date objects into strings
-        stringified_dates = [str(date) for date in dates][:-1]
-
-        # Get the relevant info for each charge type as a list
-        storage_charges, compute_charges, egress_charges = zip(*charges)
-
-        # Calculate a date's charges minus previous date
-        storage_charge_diff = self.calculate_diffs(storage_charges)
-        compute_charge_diff = self.calculate_diffs(compute_charges)
-        egress_charge_diff = self.calculate_diffs(egress_charges)
-
-        # Turn this into df to use for DataTables
-        daily_charge_df = pd.DataFrame(
-            {
-                'Date': stringified_dates,
-                'Storage charges ($)': storage_charge_diff,
-                'Compute charges ($)': compute_charge_diff,
-                'Egress charges ($)': egress_charge_diff
-            }
-        )
-
-        daily_df = daily_charge_df.to_html(
-            index=False,
-            classes = 'table table-striped" id = "dailytable',
-            justify='left'
-        )
-
-        # Add unified hover label so all charges shown
-        fig = go.Figure(
-            layout={
-                'hovermode': 'x unified'
-            }
-        )
-
-        fig.add_trace(go.Bar(
-            x=stringified_dates,
-            y=compute_charge_diff,
-            name='Compute',
-            hovertemplate='%{y:.2f}'
-            )
-        )
-
-        fig.add_trace(go.Bar(
-            x=stringified_dates,
-            y=storage_charge_diff,
-            name='Storage',
-            hovertemplate='%{y:.2f}'
-            )
-        )
-
-        fig.add_trace(go.Bar(
-            x=stringified_dates,
-            y=egress_charge_diff,
-            name='Egress',
-            hovertemplate='%{y:.2f}'
-            )
-        )
-
-        fig.update_layout(
-            title={
-                'text': "Daily Running Charges",
-                'xanchor': 'center',
-                'x': 0.5,
-                'font_size': 20
-            },
-            xaxis_title="Date",
-            xaxis_tickformat='%d %b %y',
-            yaxis_title="Daily estimated charge ($)",
-            yaxis_tickformat=",d",
-            barmode='stack',
-            width=1200,
-            height=600,
-            font_family='Helvetica',
-        )
-
-        # Add black border to bars
-        fig.update_traces(
-            marker_line_color = 'rgb(0,0,0)',
-            marker_line_width = 1
-        )
-
-        return fig, daily_df
-
-    def monthly_between_dates(self, start_month, end_month):
-        """
-        Set context for the monthly graph between start_month and end_month
-        Parameters
-        ----------
-        start_month :  str or datetime.date object
-            date in YYY-MM-DD format e.g. "2022-05-01" as first date in range
-        end_month : str or datetime.date object
-            date in YYY-MM-DD format e.g. "2022-06-01" as last date in range
-
-        Returns
-        -------
-        chart : Plotly figure object converted to HTML
-        """
-        # Filter between start of start_month
-        # And 1st of the montht hat comes after end_month
-        monthly_charges = self.totals.filter(
-            date__date__range=[
-                start_month, end_month
-            ]
-        ).values_list(
-            'date__date__month',
-            'date__date__year',
-            'storage_charges',
-            'compute_charges',
-            'egress_charges'
-        )
-
-        # Convert the end month from e.g. '2022-07-01' to '07-2022'
-        # To check if it exists later
-        check_end_month = datetime.strftime(end_month, "%m-%Y")
-
-        storage_dic = defaultdict(list)
-        compute_dic = defaultdict(list)
-        egress_dic = defaultdict(list)
-        months = []
-        storage_charges = []
-        compute_charges = []
-        egress_charges = []
-
-        # Make each a defaultdict e.g. {'5-2022': [500.20, 100.40]}
-        for item in monthly_charges:
-            storage_dic[f"0{item[0]}-{item[1]}"].append(item[2])
-            compute_dic[f"0{item[0]}-{item[1]}"].append(item[3])
-            egress_dic[f"0{item[0]}-{item[1]}"].append(item[4])
-
-        # Get the keys and sort by month-year
-        # Only need to do this for storage_dic as all have same keys
-        key_list = sorted(
-            storage_dic.keys(),
-            key = lambda x: datetime.strptime(x, '%m-%Y')
-        )
-
-        # If the end month is for the 1st of next month
-        # The month won't exist as a key in the dict
-        # Instead take the last entry of the current month
-        # To act as the 1st of the next month in its place
-        if check_end_month not in key_list:
-            storage_dic.update({
-                check_end_month: [storage_dic[key_list[-1]][-1]]
-            })
-            compute_dic.update({
-                check_end_month: [compute_dic[key_list[-1]][-1]]
-            })
-            egress_dic.update({
-                check_end_month: [egress_dic[key_list[-1]][-1]]
-            })
-
-        # Get the keys (months as e.g. '05-2022') again as list
-        # As may include next month now
-        key_list = sorted(
-            storage_dic.keys(),
-            key = lambda x: datetime.strptime(x, '%m-%Y')
-        )
-
-        # Append the first charge of each month to lists
-        for idx, month in enumerate(key_list):
-            if idx+1 <= len(key_list):
-                months.append(month)
-                storage_charges.append(storage_dic[month][0])
-                compute_charges.append(compute_dic[month][0])
-                egress_charges.append(egress_dic[month][0])
-
-        # Calculate the charge differences between months
-        storage_charges = self.calculate_diffs(storage_charges)
-        compute_charges = self.calculate_diffs(compute_charges)
-        egress_charges = self.calculate_diffs(egress_charges)
-
-        # Remove the last month from the month categories
-        # That we are taking the 1st date from but not using
-        months = months[:-1]
-
-        # Convert months to strings e.g. "May 2022" for plotting
-        converted_months = [
-            (datetime.strptime(month, "%m-%Y").strftime('%b %Y'))
-            for month in months
-        ]
-
-        # Turn this into df to use for DataTables
-        monthly_charge_df = pd.DataFrame(
-            {
-                'Month': converted_months,
-                'Storage charges': storage_charges,
-                'Compute charges': compute_charges,
-                'Egress charges': egress_charges
-            }
-        )
-
-        monthly_df = monthly_charge_df.to_html(
-            index=False,
-            #classes='table table-striped',
-            classes = 'table table-striped" id = "monthlytable',
-            justify='left'
-        )
-
-        fig = go.Figure()
-
-        fig.add_trace(go.Bar(
-            x=converted_months,
-            y=compute_charges,
-            name='Compute',
-            hovertemplate='<br>Month: %{x}<br>Charge: $%{y:.2f}<br>'
-                '<extra></extra>',
-            marker=dict(color='#636EFA')
-            )
-        )
-
-        fig.add_trace(go.Bar(
-            x=converted_months,
-            y=storage_charges,
-            name='Storage',
-            hovertemplate='<br>Month: %{x}<br>Charge: $%{y:.2f}<br>'
-                '<extra></extra>',
-            marker=dict(color='#EF553B')
-            )
-        )
-
-        fig.add_trace(go.Bar(
-            x=converted_months,
-            y=egress_charges,
-            name='Egress',
-            hovertemplate='<br>Month: %{x}<br>Charge: $%{y:.2f}<br>'
-                '<extra></extra>',
-            marker=dict(color="#00CC96")
-            )
-        )
-
-        fig.update_layout(
-            title={
-                'text': "Monthly Running Charges",
-                'xanchor': 'center',
-                'x': 0.5,
-                'font_size': 20
-            },
-            xaxis_title="Month",
-            xaxis_tickformat='%d %b %y',
-            yaxis_title="Monthly estimated charge ($)",
-            yaxis_tickformat=",d",
-            width=1200,
-            height=600,
-            font_family='Helvetica',
-        )
-
-        # Add black border to bars
-        fig.update_traces(
-            marker_line_color = 'rgb(0,0,0)',
-            marker_line_width = 1,
-        )
-
-        chart = fig.to_html()
-
-        return chart, monthly_df
-
 
 class StoragePlotFunctions():
     """Class for all of the storage plotting functions"""
@@ -563,7 +242,7 @@ class StoragePlotFunctions():
 
         exploded.rename(
             columns={
-                "name": "Type",
+                "name": "Scope",
                 "stack": "State",
                 'data': 'Monthly storage cost ($)'
             },
@@ -723,7 +402,7 @@ class StoragePlotFunctions():
 
             # Get bar colour from dict or iterate over project_colours
             live_data = {
-                'name': proj_type,
+                'name': f"{proj_type}*",
                 'data': list(
                     cost_list.values_list(
                         'Live', flat=True
@@ -736,7 +415,7 @@ class StoragePlotFunctions():
             }
 
             archived_data = {
-                'name': proj_type,
+                'name': f"{proj_type}*",
                 'data': list(
                     cost_list.values_list(
                         'Archived', flat=True
@@ -809,7 +488,7 @@ class StoragePlotFunctions():
             string_months = self.get_month_years_as_str(cost_list)
 
             live_data = {
-                'name': assay_type,
+                'name': f"*{assay_type}",
                 'data': list(
                     cost_list.values_list(
                         'Live', flat=True
@@ -822,7 +501,7 @@ class StoragePlotFunctions():
             }
 
             archived_data = {
-                'name': assay_type,
+                'name': f"*{assay_type}",
                 'data': list(
                     cost_list.values_list(
                         'Archived', flat=True
@@ -994,667 +673,3 @@ class StoragePlotFunctions():
         }
 
         return context
-
-class FilePlotFunctions():
-    """Functions for the file type storage plots"""
-
-    def __init__(self) -> None:
-        self.today_date = StoragePlotFunctions().today_date
-        self.file_type_colours = px.colors.qualitative.Pastel
-        # Chart data which is shared by all plots
-        self.chart_data = StoragePlotFunctions().chart_data
-        self.file_type_objs = FileTypeDate.objects.all()
-        self.proj_colour_dict = StoragePlotFunctions().proj_colour_dict
-        self.assay_colour_dict = StoragePlotFunctions().assay_colour_dict
-        self.project_colours = StoragePlotFunctions().project_colours
-        self.assay_colours = StoragePlotFunctions().assay_colours
-        self.file_type_categories = ['VCF', 'FASTQ', 'BAM']
-
-    def convert_to_df(self, category_chart_data, size_or_count, multi_or_all):
-        """
-        Convert chart data to a pandas df then convert it to HTML
-        So it can be shown below the graph and be easily exported
-
-        Parameters
-        ----------
-        category_chart_data : dict
-            dictionary which has all the chart attributes and data
-        size_or_count : str
-            a string to say whether to make a df for size or count
-        multi_or_all : str
-            a string to say whether all projects or looking at specific types
-        Returns
-        -------
-        chart_data : pd.DataFrame as HTML table
-            the dataframe with Date, Type, State and Total Size
-        """
-        series_data = category_chart_data['series'].copy()
-        exploded = pd.json_normalize(data=series_data).explode('data')
-
-        if size_or_count == 'size':
-            column_title = 'Total Size (GiB)'
-        else:
-            column_title = 'Total Count'
-
-        if multi_or_all == 'all':
-            scope = category_chart_data['xAxis']['categories'].copy()
-
-            # If data exists, expand the dates table according to the df length
-            # So the correct date can be added to the right row
-            if scope:
-                scope = scope * (int(len(exploded) / len(scope)))
-                exploded['Scope'] = scope
-            else:
-                scope = []
-
-            # Re-order columns
-            exploded = exploded.reindex(
-                columns=[
-                    'Scope', 'name', 'stack', 'data'
-                ]
-            )
-
-            exploded.rename(
-                columns={
-                    "name": "File Type",
-                    "stack": "State",
-                    'data': column_title
-                },
-                inplace = True
-            )
-        elif multi_or_all == 'multi':
-            # If looking at multiple diff proj / assay types
-            file_types = category_chart_data['xAxis']['categories'].copy()
-            if file_types:
-                file_types = file_types * (int(len(exploded) / len(file_types)))
-                exploded['File Type'] = file_types
-            else:
-                file_types = []
-
-            exploded = exploded.reindex(
-                columns=[
-                    'name', 'File Type', 'stack', 'data'
-                ]
-            )
-
-            exploded.rename(
-                columns={
-                    "name": "Scope",
-                    "stack": "State",
-                    'data': column_title
-                },
-                inplace = True
-            )
-
-        # Convert to HTML to easily show with DataTables
-        chart_data = exploded.to_html(
-            index=False,
-            classes='table table-striped"',
-            justify='left'
-        )
-
-        return chart_data
-
-    def todays_file_types_size_all_projects(self):
-        """
-        Returns the chart data and a df for today's file sizes in DNAnexus
-        For all of the projects aggregated
-
-        Parameters
-        ----------
-        none
-
-        Returns
-        -------
-        category_chart_data : dict
-            size series data and chart options to pass to Highcharts
-        chart_df : pd.DataFrame
-            dataframe of the series size values to show under chart
-        """
-        file_type_names = ['bam', 'fastq', 'vcf']
-
-        category_data_source = []
-        count = -2
-        for file_type in file_type_names:
-            count += 2
-            file_type_size = FileTypeDate.objects.filter(
-                date__date=self.today_date,
-                file_state__file_type__file_type=file_type
-            ).aggregate(
-                Live_Size=Sum('file_state__file_size_live'),
-                Archived_Size=Sum('file_state__file_size_archived')
-            )
-
-            live_size = file_type_size.get('Live_Size')
-            if live_size:
-                live_size = [live_size / (2**30)]
-            else:
-                live_size = []
-
-            live_data = {
-                'name': file_type.upper(),
-                'data': live_size,
-                'stack': 'Live',
-                'color': self.file_type_colours[count]
-            }
-
-            archived_size = file_type_size.get('Archived_Size')
-            if archived_size:
-                archived_size = [archived_size / (2**30)]
-            else:
-                archived_size = []
-
-            archived_data = {
-                'name': file_type.upper(),
-                'data': archived_size,
-                'stack': 'Archived',
-                'linkedTo': ':previous',
-                'color': self.file_type_colours[count],
-                'opacity': 0.8
-            }
-
-            category_data_source.append(live_data)
-            category_data_source.append(archived_data)
-
-        category_chart_data = self.chart_data.copy()
-        category_chart_data['xAxis']['categories'] = ["All projects"]
-        category_chart_data['series'] = category_data_source
-        category_chart_data['title']['text'] = (
-            f"Today's File Type Sizes - {self.today_date}"
-        )
-        category_chart_data['yAxis']['title']['text'] = "Total size (GiB)"
-        category_chart_data['tooltip']['pointFormat'] = (
-            "{series.name}: <b>{point.y:.2f} GiB </b><br>"
-            "{series.options.stack}<br>"
-        )
-
-        chart_df = self.convert_to_df(category_chart_data, "size", "all")
-
-        return category_chart_data, chart_df
-
-    def todays_file_types_count_all_projects(self):
-        """
-        Returns the chart data and a df for today's file counts in DNAnexus
-        For all of the projects aggregated
-
-        Parameters
-        ----------
-        none
-
-        Returns
-        -------
-        category_chart_data : dict
-            count series data and chart options to pass to Highcharts
-        chart_df : pd.DataFrame
-            dataframe of the series count values to show under chart
-        """
-        file_type_names = ['bam','fastq','vcf']
-
-        category_data_source = []
-        count = -2
-        for file_type in file_type_names:
-            count += 2
-            file_type_size = FileTypeDate.objects.filter(
-                date__date=self.today_date,
-                file_state__file_type__file_type=file_type
-            ).aggregate(
-                Live_Count=Sum('file_state__file_count_live'),
-                Archived_Count=Sum('file_state__file_count_archived')
-            )
-
-            live_count = file_type_size.get('Live_Count')
-            if live_count:
-                live_count = [live_count]
-            else:
-                live_count = []
-
-            live_data = {
-                'name': file_type.upper(),
-                'data': live_count,
-                'stack': 'Live',
-                'color': self.file_type_colours[count]
-            }
-
-            archived_count = file_type_size.get('Archived_Count')
-            if archived_count:
-                archived_count = [archived_count]
-            else:
-                archived_count = []
-
-            archived_data = {
-                'name': file_type.upper(),
-                'data': archived_count,
-                'stack': 'Archived',
-                'linkedTo': ':previous',
-                'color': self.file_type_colours[count],
-                'opacity': 0.8
-            }
-
-            category_data_source.append(live_data)
-            category_data_source.append(archived_data)
-
-        category_chart_data = self.chart_data.copy()
-        category_chart_data['xAxis']['categories'] = ["All projects"]
-        category_chart_data['series'] = category_data_source
-        category_chart_data['title']['text'] = (
-            f"Today's File Type Counts - {self.today_date}"
-        )
-        category_chart_data['yAxis']['title']['text'] = "Total count"
-        category_chart_data['tooltip']['pointFormat'] = (
-            "{series.name}: <b>{point.y:.f}</b><br>"
-            "{series.options.stack}<br>"
-        )
-
-        chart_df = self.convert_to_df(category_chart_data, "count", "all")
-
-        return category_chart_data, chart_df
-
-    def todays_file_types_count_project_types(self, proj_types):
-        """DO DOCSTRING"""
-        file_type_categories_dups = []
-        category_data_source = []
-        count = -1
-        for proj_type in proj_types:
-            count += 1
-            file_count_list = self.file_type_objs.filter(
-                date__date=self.today_date,
-                project__name__startswith=proj_type,
-            ).values(
-                'file_state__file_type__file_type',
-                ).annotate(
-                    Live_Count=Sum('file_state__file_count_live'),
-                    Archived_Count=Sum('file_state__file_count_archived')
-            )
-
-            file_types = [
-                entry.get('file_state__file_type__file_type').upper()
-                for entry in file_count_list
-            ]
-
-            live_data = {
-                'name': proj_type,
-                'data': [
-                    entry.get('Live_Count') for entry in file_count_list
-                ],
-                'stack': 'Live',
-                'color': self.proj_colour_dict.get(
-                    proj_type, self.project_colours[count]
-                )
-            }
-
-            archived_data = {
-                'name': proj_type,
-                'data': [
-                    entry.get('Archived_Count') for entry in file_count_list
-                ],
-                'stack': 'Archived',
-                'linkedTo': ':previous',
-                'color': self.proj_colour_dict.get(
-                    proj_type, self.project_colours[count]
-                ),
-                'opacity': 0.8
-            }
-
-            category_data_source.append(live_data)
-            category_data_source.append(archived_data)
-            file_type_categories_dups.append(file_types)
-
-        file_type_categories = file_type_categories_dups[0]
-
-        category_chart_data = self.chart_data.copy()
-        category_chart_data['xAxis']['categories'] = file_type_categories
-        category_chart_data['series'] = category_data_source
-        category_chart_data['title']['text'] = (
-            f"Today's File Counts By Project Type - {self.today_date}"
-        )
-        category_chart_data['yAxis']['title']['text'] = "Total Count"
-        category_chart_data['tooltip']['pointFormat'] = (
-            "{series.name}: <b>{point.y:.0f}</b><br>"
-            "{series.options.stack}<br>"
-        )
-
-        chart_df = self.convert_to_df(category_chart_data, "count", "multi")
-
-        return category_chart_data, chart_df
-
-    def todays_file_types_size_project_types(self, proj_types):
-        """DO DOCSTRING"""
-        file_type_categories_dups = []
-        category_data_source = []
-        count = -1
-        for proj_type in proj_types:
-            count += 1
-            file_size_list = self.file_type_objs.filter(
-                date__date=self.today_date,
-                project__name__startswith=proj_type,
-            ).values(
-                'file_state__file_type__file_type',
-                ).annotate(
-                    Live_Size=Sum('file_state__file_size_live'),
-                    Archived_Size=Sum('file_state__file_size_archived')
-            )
-
-            file_types = [
-                entry.get('file_state__file_type__file_type').upper()
-                for entry in file_size_list
-            ]
-
-            live_data = {
-                'name': proj_type,
-                'data': [
-                    (entry.get('Live_Size')/(2**30))
-                    for entry in file_size_list
-                ],
-                'stack': 'Live',
-                'color': self.proj_colour_dict.get(
-                    proj_type, self.project_colours[count]
-                )
-            }
-
-            archived_data = {
-                'name': proj_type,
-                'data': [
-                    (entry.get('Archived_Size')/(2**30))
-                    for entry in file_size_list
-                ],
-                'stack': 'Archived',
-                'linkedTo': ':previous',
-                'color': self.proj_colour_dict.get(
-                    proj_type, self.project_colours[count]
-                ),
-                'opacity': 0.8
-            }
-
-            category_data_source.append(live_data)
-            category_data_source.append(archived_data)
-            file_type_categories_dups.append(file_types)
-
-        file_type_categories = file_type_categories_dups[0]
-
-        category_chart_data = self.chart_data.copy()
-        category_chart_data['xAxis']['categories'] = file_type_categories
-        category_chart_data['series'] = category_data_source
-        category_chart_data['title']['text'] = (
-            f"Today's File Sizes By Project Type - {self.today_date}"
-        )
-        category_chart_data['yAxis']['title']['text'] = "Total Size"
-        category_chart_data['tooltip']['pointFormat'] = (
-            "{series.name}: <b>{point.y:.2f} GiB </b><br>"
-            "{series.options.stack}<br>"
-        )
-
-        chart_df = self.convert_to_df(category_chart_data, "size", "multi")
-
-        return category_chart_data, chart_df
-
-    def todays_file_types_count_assay_types(self, assay_types):
-        """DO DOCSTRING"""
-        file_type_categories_dups = []
-        category_data_source = []
-        count = -1
-        for assay_type in assay_types:
-            count += 1
-            file_count_list = self.file_type_objs.filter(
-                date__date=self.today_date,
-                project__name__endswith=assay_type,
-            ).values(
-                'file_state__file_type__file_type',
-            ).annotate(
-                    Live_Count=Sum('file_state__file_count_live'),
-                    Archived_Count=Sum('file_state__file_count_archived')
-            )
-
-            file_types = [
-                entry.get('file_state__file_type__file_type').upper()
-                for entry in file_count_list
-            ]
-
-            live_data = {
-                'name': assay_type,
-                'data': [
-                    entry.get('Live_Count')
-                    for entry in file_count_list
-                ],
-                'stack': 'Live',
-                'color': self.assay_colour_dict.get(
-                    assay_type, self.assay_colours[count]
-                )
-            }
-
-            archived_data = {
-                'name': assay_type,
-                'data': [
-                    entry.get('Archived_Count')
-                    for entry in file_count_list
-                ],
-                'stack': 'Archived',
-                'linkedTo': ':previous',
-                'color': self.assay_colour_dict.get(
-                    assay_type, self.assay_colours[count]
-                ),
-                'opacity': 0.8
-            }
-
-            category_data_source.append(live_data)
-            category_data_source.append(archived_data)
-            file_type_categories_dups.append(file_types)
-
-        file_type_categories = file_type_categories_dups[0]
-
-        category_chart_data = self.chart_data.copy()
-        category_chart_data['xAxis']['categories'] = file_type_categories
-        category_chart_data['series'] = category_data_source
-        category_chart_data['title']['text'] = (
-            f"Today's File Counts By Assay Type - {self.today_date}"
-        )
-        category_chart_data['yAxis']['title']['text'] = "Total Count"
-        category_chart_data['tooltip']['pointFormat'] = (
-            "{series.name}: <b>{point.y:.0f} </b><br>"
-            "{series.options.stack}<br>"
-        )
-
-        chart_df = self.convert_to_df(category_chart_data, "count", "multi")
-
-        return category_chart_data, chart_df
-
-    def todays_file_types_size_assay_types(self, assay_types):
-        """DO DOCSTRING"""
-        file_type_categories_dups = []
-        category_data_source = []
-        count = -1
-        for assay_type in assay_types:
-            count += 1
-            file_size_list = self.file_type_objs.filter(
-                date__date=self.today_date,
-                project__name__endswith=assay_type,
-            ).values(
-                'file_state__file_type__file_type',
-            ).annotate(
-                    Live_Size=Sum('file_state__file_size_live'),
-                    Archived_Size=Sum('file_state__file_size_archived')
-            )
-
-            file_types = [
-                entry.get('file_state__file_type__file_type').upper()
-                for entry in file_size_list
-            ]
-
-            live_data = {
-                'name': assay_type,
-                'data': [
-                    (entry.get('Live_Size')/(2**30))
-                    for entry in file_size_list
-                ],
-                'stack': 'Live',
-                'color': self.assay_colour_dict.get(
-                    assay_type, self.assay_colours[count]
-                )
-            }
-
-            archived_data = {
-                'name': assay_type,
-                'data': [
-                    (entry.get('Archived_Size')/(2**30))
-                    for entry in file_size_list
-                ],
-                'stack': 'Archived',
-                'linkedTo': ':previous',
-                'color': self.assay_colour_dict.get(
-                    assay_type, self.assay_colours[count]
-                ),
-                'opacity': 0.8
-            }
-
-            category_data_source.append(live_data)
-            category_data_source.append(archived_data)
-            file_type_categories_dups.append(file_types)
-
-        file_type_categories = file_type_categories_dups[0]
-
-        category_chart_data = self.chart_data.copy()
-        category_chart_data['xAxis']['categories'] = file_type_categories
-        category_chart_data['series'] = category_data_source
-        category_chart_data['title']['text'] = (
-            f"Today's File Sizes By Assay Type - {self.today_date}"
-        )
-        category_chart_data['yAxis']['title']['text'] = "Total Size (GiB)"
-        category_chart_data['tooltip']['pointFormat'] = (
-            "{series.name}: <b>{point.y:.2f} GiB</b><br>"
-            "{series.options.stack}<br>"
-        )
-
-        chart_df = self.convert_to_df(category_chart_data, "size", "multi")
-
-        return category_chart_data, chart_df
-
-    def todays_file_types_count_assay_and_proj_types(
-        self, project_type, assay_type
-    ):
-        """DO DOCSTRING"""
-        category_data_source = []
-        file_count_list = self.file_type_objs.filter(
-            date__date=self.today_date,
-            project__name__startswith=project_type,
-            project__name__endswith=assay_type
-        ).values(
-            'file_state__file_type__file_type',
-        ).annotate(
-                Live_Count=Sum('file_state__file_count_live'),
-                Archived_Count=Sum('file_state__file_count_archived')
-        )
-
-        file_type_categories = [
-            entry.get('file_state__file_type__file_type').upper()
-            for entry in file_count_list
-        ]
-
-        live_data = {
-            'name': f"{project_type}*{assay_type}",
-            'data': [
-                entry.get('Live_Count')
-                for entry in file_count_list
-            ],
-            'stack': 'Live',
-            'color': self.proj_colour_dict.get(
-                project_type, self.project_colours[0]
-            ),
-        }
-
-        archived_data = {
-            'name': f"{project_type}*{assay_type}",
-            'data': [
-                entry.get('Archived_Count')
-                for entry in file_count_list
-            ],
-            'stack': 'Archived',
-            'linkedTo': ':previous',
-            'color': self.proj_colour_dict.get(
-                project_type, self.project_colours[0]
-            ),
-            'opacity': 0.8
-        }
-
-        category_data_source.append(live_data)
-        category_data_source.append(archived_data)
-
-        category_chart_data = self.chart_data.copy()
-        category_chart_data['xAxis']['categories'] = file_type_categories
-        category_chart_data['series'] = category_data_source
-        category_chart_data['title']['text'] = (
-            f"Today's File Counts By Project + Assay Type - {self.today_date}"
-        )
-        category_chart_data['yAxis']['title']['text'] = "Total Count"
-        category_chart_data['tooltip']['pointFormat'] = (
-            "{series.name}: <b>{point.y:.0f}</b><br>"
-            "{series.options.stack}<br>"
-        )
-
-        chart_df = self.convert_to_df(category_chart_data, "count", "multi")
-
-        return category_chart_data, chart_df
-
-    def todays_file_types_size_assay_and_proj_types(
-        self, project_type, assay_type
-    ):
-        """DO DOCSTRING"""
-        category_data_source = []
-        file_size_list = self.file_type_objs.filter(
-            date__date=self.today_date,
-            project__name__startswith=project_type,
-            project__name__endswith=assay_type
-        ).values(
-            'file_state__file_type__file_type',
-        ).annotate(
-                Live_Size=Sum('file_state__file_size_live'),
-                Archived_Size=Sum('file_state__file_size_archived')
-        )
-
-        file_type_categories = [
-            entry.get('file_state__file_type__file_type').upper()
-            for entry in file_size_list
-        ]
-
-        live_data = {
-            'name': f"{project_type}*{assay_type}",
-            'data': [
-                (entry.get('Live_Size')/(2**30))
-                for entry in file_size_list
-            ],
-            'stack': 'Live',
-            'color': self.proj_colour_dict.get(
-                project_type, self.project_colours[0]
-            ),
-        }
-
-        archived_data = {
-            'name': f"{project_type}*{assay_type}",
-            'data': [
-                (entry.get('Archived_Size')/(2**30))
-                for entry in file_size_list
-            ],
-            'stack': 'Archived',
-            'linkedTo': ':previous',
-            'color': self.proj_colour_dict.get(
-                project_type, self.project_colours[0]
-            ),
-            'opacity': 0.8
-        }
-
-        category_data_source.append(live_data)
-        category_data_source.append(archived_data)
-
-        category_chart_data = self.chart_data.copy()
-        category_chart_data['xAxis']['categories'] = file_type_categories
-        category_chart_data['series'] = category_data_source
-        category_chart_data['title']['text'] = (
-            f"Today's File Sizes By Project + Assay Type - {self.today_date}"
-        )
-        category_chart_data['yAxis']['title']['text'] = "Total Size (GiB)"
-        category_chart_data['tooltip']['pointFormat'] = (
-            "{series.name}: <b>{point.y:.2f} GiB</b><br>"
-            "{series.options.stack}<br>"
-        )
-
-        chart_df = self.convert_to_df(category_chart_data, "size", "multi")
-
-        return category_chart_data, chart_df
